Update test filenames to match module names.
