+++ pyface/tasks/tests/test_task_gui_application.py
