+++ pyface/tests/test_gui_application.py
@@ -5,12 +5,16 @@ import os
 from shutil import rmtree
 import sys
 from tempfile import mkdtemp
-from unittest import TestCase
+import unittest
 
 from traits.api import Bool
-from traits.testing.unittest_tools import UnittestTools
 
-from ..application import ApplicationExit, BaseApplication
+from ..gui_application import Application
+from ..application_window import ApplicationWindow
+from ..toolkit import toolkit_object
+
+GuiTestAssistant = toolkit_object('util.gui_test_assistant:GuiTestAssistant')
+no_gui_test_assistant = (GuiTestAssistant.__name__ == 'Unimplemented')
 
 
 EVENTS = [
@@ -18,7 +22,7 @@ EVENTS = [
 ]
 
 
-class TestingApp(BaseApplication):
+class TestingApp(Application):
 
     #: Whether the app should start cleanly.
     start_cleanly = Bool(True)
@@ -38,6 +42,9 @@ class TestingApp(BaseApplication):
     #: Whether to veto a call to the exit method.
     veto_exit = Bool(False)
 
+    #: Whether to veto a closing a window.
+    veto_close = Bool(False)
+
     #: Whether or not a call to the exit method was vetoed.
     exit_vetoed = Bool(False)
 
@@ -49,35 +56,46 @@ class TestingApp(BaseApplication):
 
     def start(self):
         super(TestingApp, self).start()
-        return self.start_cleanly
+        if not self.start_cleanly:
+            return False
+        window = ApplicationWindow()
+        window.on_trait_change(
+            lambda event: setattr(event, 'veto', self.veto_close), 'closing'
+        )
+        self.add_window(window)
+        return True
 
     def stop(self):
         super(TestingApp, self).stop()
         return self.stop_cleanly
 
-    def _run(self):
-        super(TestingApp, self)._run()
-        if self.do_exit:
-            if self.error_exit:
-                raise ApplicationExit("error message")
-            else:
-                self.exit(self.force_exit)
-            self.exit_vetoed = True
-        return True
+    def _on_window_closing(self, window, trait, old, new):
+        if self.veto_close:
+            new.veto = True
 
     def _exiting_fired(self, event):
         event.veto = self.veto_exit
 
     def _prepare_exit(self):
+        super(TestingApp, self)._prepare_exit()
         self.exit_prepared = True
         if self.exit_prepared_error:
             raise Exception("Exit preparation failed")
 
 
-class TestBaseApplication(TestCase, UnittestTools):
+@unittest.skipIf(no_gui_test_assistant, 'No GuiTestAssistant')
+class TestApplication(unittest.TestCase, GuiTestAssistant):
 
     def setUp(self):
+        GuiTestAssistant.setUp(self)
         self.application_events = []
+
+        if toolkit_object.toolkit == 'wx':
+            import wx
+            self.event_loop()
+            wx.GetApp().DeletePendingEvents()
+        else:
+            self.event_loop()
 
     def event_listener(self, event):
         self.application_events.append(event)
@@ -89,7 +107,7 @@ class TestBaseApplication(TestCase, UnittestTools):
     def test_defaults(self):
         from traits.etsconfig.api import ETSConfig
 
-        app = BaseApplication()
+        app = Application()
 
         self.assertEqual(app.home, ETSConfig.application_home)
         self.assertEqual(app.user_data, ETSConfig.user_data)
@@ -98,7 +116,7 @@ class TestBaseApplication(TestCase, UnittestTools):
     def test_logging_setup(self):
         dirpath = mkdtemp()
         home = os.path.join(dirpath, "test")
-        app = BaseApplication(home=home)
+        app = Application(home=home)
 
         app.setup_logging()
         try:
@@ -113,7 +131,7 @@ class TestBaseApplication(TestCase, UnittestTools):
 
     def test_excepthook(self):
         excepthook = sys.excepthook
-        app = BaseApplication()
+        app = Application()
 
         app.install_excepthook()
         try:
@@ -126,7 +144,7 @@ class TestBaseApplication(TestCase, UnittestTools):
     def test_initialize_application_home(self):
         dirpath = mkdtemp()
         home = os.path.join(dirpath, "test")
-        app = BaseApplication(home=home)
+        app = Application(home=home)
 
         app.initialize_application_home()
         try:
@@ -135,34 +153,27 @@ class TestBaseApplication(TestCase, UnittestTools):
             rmtree(dirpath)
 
     def test_lifecycle(self):
-        app = BaseApplication()
+
+        app = Application()
         self.connect_listeners(app)
+        window = ApplicationWindow()
+        app.on_trait_change(lambda: app.add_window(window), 'started')
 
         with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit)
             result = app.run()
 
         self.assertTrue(result)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS)
-
-    def test_exit(self):
-        app = TestingApp(do_exit=True)
-        self.connect_listeners(app)
-
-        with self.assertMultiTraitChanges([app], EVENTS, []):
-            result = app.run()
-
-        self.assertTrue(result)
-        self.assertFalse(app.exit_vetoed)
-        self.assertTrue(app.exit_prepared)
-        event_order = [event.event_type for event in self.application_events]
-        self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
 
     def test_exit_prepare_error(self):
-        app = TestingApp(do_exit=True, exit_prepared_error=True)
+        app = TestingApp(exit_prepared_error=True)
         self.connect_listeners(app)
 
         with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit)
             result = app.run()
 
         self.assertTrue(result)
@@ -170,12 +181,24 @@ class TestBaseApplication(TestCase, UnittestTools):
         self.assertTrue(app.exit_prepared)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
 
     def test_veto_exit(self):
-        app = TestingApp(do_exit=True, veto_exit=True)
+        app = TestingApp(veto_exit=True)
         self.connect_listeners(app)
 
+        def hard_exit():
+            app.exit_vetoed = True
+            for window in app.windows:
+                window.destroy()
+                print( window.control)
+            app.windows = []
+            self.event_loop()
+            self.gui.stop_event_loop()
+
         with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit)
+            self.gui.invoke_after(200, hard_exit)
             result = app.run()
 
         self.assertTrue(result)
@@ -183,12 +206,38 @@ class TestBaseApplication(TestCase, UnittestTools):
         self.assertFalse(app.exit_prepared)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
+
+    def test_veto_close(self):
+        app = TestingApp(veto_close=True)
+        self.connect_listeners(app)
+
+        def hard_exit():
+            app.exit_vetoed = True
+            for window in app.windows:
+                window.destroy()
+            app.windows = []
+            self.event_loop()
+            self.gui.stop_event_loop()
+
+        with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit)
+            self.gui.invoke_after(200, hard_exit)
+            result = app.run()
+
+        self.assertTrue(result)
+        self.assertTrue(app.exit_vetoed)
+        self.assertFalse(app.exit_prepared)
+        event_order = [event.event_type for event in self.application_events]
+        self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
 
     def test_force_exit(self):
         app = TestingApp(do_exit=True, force_exit=True, veto_exit=True)
         self.connect_listeners(app)
 
         with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit, True)
             result = app.run()
 
         self.assertTrue(result)
@@ -196,18 +245,22 @@ class TestBaseApplication(TestCase, UnittestTools):
         self.assertTrue(app.exit_prepared)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
 
-    def test_error_exit(self):
-        app = TestingApp(do_exit=True, error_exit=True)
+    def test_force_exit_close_veto(self):
+        app = TestingApp(do_exit=True, force_exit=True, veto_close=True)
         self.connect_listeners(app)
 
         with self.assertMultiTraitChanges([app], EVENTS, []):
+            self.gui.invoke_after(100, app.exit, True)
             result = app.run()
 
-        self.assertFalse(result)
+        self.assertTrue(result)
         self.assertFalse(app.exit_vetoed)
+        self.assertTrue(app.exit_prepared)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS)
+        self.assertEqual(app.windows, [])
 
     def test_bad_start(self):
         app = TestingApp(start_cleanly=False)
@@ -219,14 +272,17 @@ class TestBaseApplication(TestCase, UnittestTools):
         self.assertFalse(result)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS[:1])
+        self.assertEqual(app.windows, [])
 
     def test_bad_stop(self):
         app = TestingApp(stop_cleanly=False)
         self.connect_listeners(app)
 
         with self.assertMultiTraitChanges([app], EVENTS[:-1], EVENTS[-1:]):
+            self.gui.invoke_after(100, app.exit, True)
             result = app.run()
 
         self.assertFalse(result)
         event_order = [event.event_type for event in self.application_events]
         self.assertEqual(event_order, EVENTS[:-1])
+        self.assertEqual(app.windows, [])
